+++ src/Interfaces/SectionInstanceInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Nadar\PhpComposerReader\Interfaces;
+
+/**
+ * A section instance interface.
+ * 
+ * The section instance interface represents a single item of a given section, for example:
+ * 
+ * + a require entry
+ * + an auto load entry
+ * 
+ * @author Basil Suter <basil@nadar.io>
+ * @since 1.4.0
+ */
+interface SectionInstanceInterface
+{
+}